feat(abbreviations): Use ISO-3166-1 alpha3 code from WOF records
Use the `wof:country_alpha3` field, if present, to set the corresponding
field on Pelias country records.

This allows us to remove the troublesome (see https://github.com/pelias/api/issues/1179)
iso3166-1 NPM package.

One implication of the change is that it does mean we are trusting the
data in WOF to be correct.